Implement dynamic value generators
This change enables values to be generated dynamically for fuzzable
and custom primitives in RESTler.

For every combination (up to 'max_combinations'), if a custom generator
is available for a fuzzable or custom primitive, RESTler will use it to
generate a new value instead of using values from the static dictionary.

The user can add value generators in python corresponding to fuzzable and
custom primitives (see documentation for the list of supported primitives).
RESTler fetches the generators from a dictionary with the same structure
as the custom values dictionary.  The generator dictionary must be defined in a
python file and specified in the engine settings (see documentation).
A template file with placeholders for all possible generators corresponding
to the dictionary is generated when compiling the spec.

**Implementation notes:**
When RESTler generates sequences of requests, the prefix of the sequence
(all requests up to the last one) contains requests that have been
previously successfully executed.  In order to send the request with
the same set of values in the presence of dynamic generators, the
fuzzable values previously sent are cached and simply plugged in
when a sequence is being re-rendered.

There was some complexity required in integrating with 'render_iter',
which to date has been stateless - if skip=N was specified, it simply
re-generated all the combinations in order to get to N.
1) The same value needs to be used for re-rendering sequences.
This is solved as noted above, by saving the previously rendered values.

2) Since it's not clear how generated values should be combined with others for a given schema and
primitive type, a simple solution was chosen to fetch a new value any time it is needed.
This ensures that all the static values will be combined as usual up to 'max_combinations',
and the necessary generated values will be plugged in as needed.
For example, if there are two fuzzable strings in the request,
and 3 'restler_fuzzable_string' values in the static dictionary,
a total of 9 combinations will be generated.
If these same 3 values are returned in a value generator instead,
only 3 combinations will be tested ((v1,v1), (v2,v2), and (v3, v3)).

3) Before testing, RESTler iterates to see how many total combinations there are.
The state of any cached values, including the value generators, must be reset after
this initial count to make sure all of the values intended to be tested are used.

Because requests are deep copied when saving rendered sequences,
special logic needed to be added to exclude the generators and associated data from
being copied.

Testing:
- Manual testing with demo_server
- New compiler baseline test for the generated dictionary
- [TODO]Engine unit test to confirm sequence re-rendering works
correctly with the same values when values are dynamically generated
- Engine test to confirm that the number of combinations is correct